requirements/base.txt
Django==1.9.7
django-environ>=0.4.0
#
# apt-get install python-dev libmysqlclient-dev
#
mysqlclient>=1.3.7

#
# apt-get install python3-dev
#
easy-thumbnails>=2.3
djangorestframework==3.3.3
markdown
django-filter
python-memcached>=1.57
django-storages>=1.4.1
boto
gunicorn
redis
hiredis
sendgrid-django
sqlparse

#
# apt-get install build-essential libffi-dev python-dev
#
bcrypt>=3.1.0
django-redis
django-q
django-picklefield
arrow
blessed
requests
matplotlib
scipy
geopy
mongoengine
django-rest-framework-mongoengine
lxml
beautifulsoup4
opencv-python
